Refactor KeyStore into utils
