+++ ingestion-beam/src/main/java/com/mozilla/telemetry/decoder/DecryptPioneerPayloads.java
@@ -3,15 +3,11 @@ package com.mozilla.telemetry.decoder;
 import com.fasterxml.jackson.databind.node.ObjectNode;
 import com.mozilla.telemetry.transforms.FailureMessage;
 import com.mozilla.telemetry.transforms.PubsubConstraints;
-import com.mozilla.telemetry.util.BeamFileInputStream;
 import com.mozilla.telemetry.util.Json;
+import com.mozilla.telemetry.util.KeyStore;
 import java.io.IOException;
-import java.io.InputStream;
-import java.io.UncheckedIOException;
 import java.security.PrivateKey;
 import java.util.Collections;
-import java.util.HashMap;
-import java.util.Map;
 import org.apache.beam.sdk.io.gcp.pubsub.PubsubMessage;
 import org.apache.beam.sdk.options.ValueProvider;
 import org.apache.beam.sdk.transforms.FlatMapElements;
@@ -21,9 +17,7 @@ import org.apache.beam.sdk.transforms.WithFailures;
 import org.apache.beam.sdk.transforms.WithFailures.Result;
 import org.apache.beam.sdk.values.PCollection;
 import org.apache.beam.sdk.values.TypeDescriptor;
-import org.apache.beam.vendor.grpc.v1p21p0.io.grpc.internal.IoUtils;
 import org.jose4j.jwe.JsonWebEncryption;
-import org.jose4j.jwk.PublicJsonWebKey;
 import org.jose4j.lang.JoseException;
 
 public class DecryptPioneerPayloads extends
@@ -32,7 +26,7 @@ public class DecryptPioneerPayloads extends
   private final ValueProvider<String> keysLocation;
   private transient KeyStore keyStore;
 
-  public static DecryptPioneerPayloads of(ValueProvider keysLocation) {
+  public static DecryptPioneerPayloads of(ValueProvider<String> keysLocation) {
     return new DecryptPioneerPayloads(keysLocation);
   }
 
@@ -55,54 +49,6 @@ public class DecryptPioneerPayloads extends
                 ee.exception());
           }
         }));
-  }
-
-  private class KeyStore {
-
-    public PrivateKey getKey(String path) {
-      ensureKeysLoaded();
-      PrivateKey key = keys.get(path);
-      return key;
-    }
-
-    private final String keysLocation;
-    private transient Map<String, PrivateKey> keys;
-
-    // enable this to read from gcs via BeamFileInputStream
-    public KeyStore(String keysLocation) {
-      this.keysLocation = keysLocation;
-    }
-
-    private void loadAllKeys() throws IOException {
-      final Map<String, PrivateKey> tempKeys = new HashMap<>();
-      InputStream inputStream;
-      try {
-        inputStream = BeamFileInputStream.open(this.keysLocation);
-      } catch (IOException e) {
-        throw new IOException("Exception thrown while reading private key");
-      }
-      byte[] serializedKeyBytes = IoUtils.toByteArray(inputStream);
-      String serializedKey = new String(serializedKeyBytes);
-
-      try {
-        PublicJsonWebKey key = PublicJsonWebKey.Factory.newPublicJwk(serializedKey);
-        tempKeys.put("*", key.getPrivateKey());
-      } catch (JoseException e) {
-        throw new RuntimeException(e);
-      }
-
-      keys = tempKeys;
-    }
-
-    private void ensureKeysLoaded() {
-      if (keys == null) {
-        try {
-          loadAllKeys();
-        } catch (IOException e) {
-          throw new UncheckedIOException("unexpected error while loading keys", e);
-        }
-      }
-    }
   }
 
   private class Fn implements ProcessFunction<PubsubMessage, Iterable<PubsubMessage>> {
